requirements.txt
asgiref==3.9.1
Django==5.2.5
django-ckeditor==6.7.3
django-extensions==4.1
django-js-asset==3.1.2
django-modeltranslation==0.19.16
pillow==11.3.0
pytz==2025.2
sqlparse==0.5.3
